requirements.txt
# Веб-фреймворк
fastapi==0.118.0
uvicorn==0.37.0

# Работа с Google Sheets
gspread==6.2.1
google-auth==2.41.0
google-auth-oauthlib==1.2.2

# Работа с данными и валидация
pydantic==2.11.9
pydantic-settings==2.11.0
sqlmodel==0.0.25

# Вспомогательные
python-dotenv==1.1.1
python-multipart==0.0.20

# HTTP клиент
httpx==0.28.1
requests==2.32.5
requests-oauthlib==2.0.0